+++ Programacion Orientada a Objetos/Clase 25/Personas/LectorArchivos.ts
@@ -2,23 +2,13 @@ import Persona from "./Persona";
 import * as FS from "fs";
 
 export default class Lector{
-    private personas: Persona[]
+    private lectorArchivos:string
 
-    public constructor(){
-        this.personas = [];
+    public constructor(lectorArchivos:string){
+        this.lectorArchivos = lectorArchivos;
 
     }
-
-    public cargarPersonas():void{
-        let personas:string[] = (FS.readFileSync('personas.txt','utf8')).split('\n');
-        let propiedadesPersona:string[] = [];
-        personas.forEach(personaString => {
-            propiedadesPersona = personaString.split(";")
-            this.personas.push(new Persona(propiedadesPersona[0],propiedadesPersona[1],propiedadesPersona[2],parseInt(propiedadesPersona[3]),parseInt(propiedadesPersona[4])))
-        });
-    }
-    public mostrarPersonas(){
-        console.log(this.personas)
-    }
-
+   public cargaDatos(lectorArchivos,separador):string{
+       return this.lectorArchivos
+   }
 }
